refactor: Auth 관련 파일 폴더 구조 정리
