+++ src/domains/course/enroll/components/BatchEnrollButton.tsx
@@ -2,7 +2,7 @@ import { coursesQuery } from '@apis/courses';
 import Button from '@components/actions/Button';
 import { useShowToast } from '@components/toast/ToastProvider';
 import { useCallback } from 'react';
-import { useCheckCourses } from '../../components/courseList/context/CheckCoursesContext';
+import { useCheckCourses } from '../../context/CheckCoursesContext';
 import { getEnrollResultMessage } from '../utils/getEnrollResultMessage';
 import styled from '@emotion/styled';
