+++ src/pages/auth/login/LoginPage.tsx
@@ -6,8 +6,8 @@ import { useNavigate } from 'react-router';
 import { ROUTE_PATH } from '@routes/routePath';
 import { useShowToast } from '@components/toast/ToastProvider';
 import { ApiError } from '@apis/apiClient';
-import { useAuth } from '@domains/auth/login/context/AuthProvider';
-import { loginFormRules } from '@domains/auth/utils/loginFormRules';
+import { useAuth } from '@domains/auth/context/AuthProvider';
+import { loginFormRules } from '@domains/auth/validation';
 
 type LoginFormData = {
   email: string;
